Multiple fixes (Details in the description) (#35)
* Allow setting of load balancer listener IP

This change depends on https://github.com/apache/libcloud/pull/818

* Use RawConfigParser instead of SafeConfigParser

This change keeps get_credentials() from returning false when
DIDATA_PASSWORD is a strong password that contains the '%' symbol.
Before this change, getting an entry with that symbol would cause
a ConfigParser.InterpolationSyntaxError which would cause the function
to return false. The Ansible user would then get a confusion "User
credentials not found" error. Since we don't expect the user to employ
any string interpolation in this file, we might as well do away with it.

* Print out a more helpful message when invalid credentials are provided.

Before this change, Ansible would only print out '' as an error message
because libcloud's InvalidCredentialsError type did not have a message
attribute. This works around that problem by explicitly writing out a
more helpful message when the above exception is caught.